Read simple features from the Manifold driver

Read a drawing from a Manifold connection.

mfd_read_db(con = NULL, table, query = "", geom_column = NULL,
  WHERE = "", ..., quiet = TRUE)

Arguments

con

connection object

table

table to read, can be a Drawing name or its child table

query

query to run, optional

geom_column

this is the name of the column as returned (not the name of the column natively)

WHERE

optional WHERE clause, as in "WHERE ID = 1 AND size < 10"

...

other arguments passed along (none currently)

quiet

keep quiet by default

Details

By default the entire drawing is read with all non-intrinsic columns. The intrinsic column "Geom (I)" is cast to WKB within Manifold and then interpreted using st_as_sfc.